requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
cryptography==36.0.2
defusedxml==0.7.1
Django==3.1.12
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-emoji-picker==0.0.6
django-extensions==3.1.5
django-phone-field==1.8.1
django-tinymce==3.4.0
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
